+++ src/FluxPwaGenerator.mjs
@@ -26,6 +26,19 @@ export class FluxPwaGenerator {
     }
 
     /**
+     * @param {string} icon_template_svg_file
+     * @param {string} manifest_json_file
+     * @returns {Promise<void>}
+     */
+    async generateIcons(icon_template_svg_file, manifest_json_file) {
+        await (await import("./Pwa/GenerateIcons.mjs")).GenerateIcons.new()
+            .generateIcons(
+                icon_template_svg_file,
+                manifest_json_file
+            );
+    }
+
+    /**
      * @param {string} index_template_html_file
      * @param {string} index_html_file
      * @param {string} manifest_json_file